+++ Project_Flight/DLL_DAL/Model/VeChuyenBayModel.cs
@@ -9,12 +9,12 @@ namespace DLL_DAL.Model
     public class VeChuyenBayModel
     {
         long Id;
-        float DonGia;
+        Double DonGia;
         long ChuyenBayId;
         long HangVeId;
 
         public long id { get => Id; set => Id = value; }
-        public float donGia { get => DonGia; set => DonGia = value; }
+        public Double donGia { get => DonGia; set => DonGia = value; }
         public long chuyenBayID { get => ChuyenBayId; set => ChuyenBayId = value; }
         public long hangVeID { get => HangVeId; set => HangVeId = value; }
     }
